paper/requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langchain-ollama>=0.2.0
langchain-mistralai>=0.2.0
python-dotenv>=1.0.0
fpdf2>=2.7.0
